Final cleanup and fixes for merging of sample apps.
