+++ examples/android/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/AppTheme"
         android:supportsRtl="false">
         <activity
-            android:name=".UserActivity"
+            android:name=".HomeActivity"
             android:exported="true"
             android:label="@string/app_name">
             <intent-filter>
@@ -41,7 +41,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".internal.OpenWithActivity"
+            android:name=".OpenWithActivity"
             android:exported="true"
             android:label="@string/title_activity_openwith">
             <intent-filter>
