+++ examples/android/src/main/res/layout/activity_home.xml
